authorization middleware & notes controller with JWT

+++ notes-api/controllers/notesController.js
@@ -2,12 +2,13 @@ import { getDBConnection } from "../database/db.js";
 
 
 
-// Notes CRUD but i didnt implement any type of user IDing
+// Notes CRUD 
 export async function getAllNotes(req,res){
     const db = await getDBConnection()
 
     const notes = await db.all(
-        `SELECT * FROM notes`
+        `SELECT * FROM notes WHERE user_ID = ?`,
+        [req.user.user_id]
     )
 
     return res.status(200).json({notes})
@@ -18,14 +19,16 @@ export async function addNote(req,res){
     const db = await getDBConnection()
     try{
         const {title,content} = req.body
-
+        const {user_id} = req.user
+        console.log(req.user)
         const result = await db.run(`
-            INSERT INTO notes(title,content) VALUES(?,?)
-            `,[title,content])
+            INSERT INTO notes(title,content,user_ID) VALUES(?,?,?)
+            `,[title,content,user_id])
         return res.status(201).json({id : result.lastID,title,content})
 
     }catch(err){
-        return res.status(500).json({"error": `failure to add note due to ${err}`})
+        console.error(err)
+        return res.status(500).json({message: `failure to add note`})
     }
     
 }
@@ -37,7 +40,7 @@ export async function deleteNote(req,res){
     try{
 
         await db.run(`
-            DELETE FROM notes WHERE id = ?`, [noteId])
+            DELETE FROM notes WHERE user_ID = ? AND id = ?`, [req.user.user_id,noteId])
         
     
         return res.status(201).json({"message": "note deleted"})
@@ -55,8 +58,8 @@ export async function updateNote(req,res){
         await db.run(`
             UPDATE notes
             SET title = ? , content = ?
-            WHERE id = ?
-            `, [title,content, noteId])
+            WHERE user_ID = ? AND id = ?
+            `, [title,content,req.user.user_id, noteId])
 
         res.status(201).json({'message': "note updated"})
     }catch(err){
